+++ changelog.md
@@ -6,15 +6,20 @@
 
 - Added Personio icon in the WordPress icon library (for WordPress >= 7.1)
 - Added new logging events during plugin activation and deactivation
+- Added more PHP Unit Tests running before each release
 
 ### Changed
 
 - Small updates for die API v2, but still not usable as Personio do not deliver the most of v1 data for positions
+- Optimized cronjob handling, which now also supports automatic cron-healing if a server-side cronjob is used
 
 ### Fixed
 
 - Fixed the statement to get our own Block templates from the database (thanks @philipp-marquardt)
 - Fixed removing of schedules with attributes during deactivation of plugin
+- Fixed a wrong default interval for the report by email
+- Fixed wrong selection of intervals for the report by email
+- Fixed potential error of icon collection has been already registered
 
 ## [5.6.0] - 18.08.2026